cautiously optimistic strategy will work, static but related to the client which is why i'm leaving it here instead of moving into utils
