+++ packages/hedera-wallet-snap/packages/snap/src/types/hedera.ts
@@ -131,7 +131,6 @@ export type TxRecord = {
 };
 
 export type HederaService = {
-
   getNodeStakingInfo(): Promise<MirrorStakingInfo[]>;
 
   getMirrorAccountInfo(idOrAliasOrEvmAddress: string): Promise<AccountInfo>;
@@ -164,63 +163,6 @@ export type SimpleHederaClient = {
 
   // get the associated account ID
   getAccountId(): AccountId;
-
-  getAccountInfo(accountId: string): Promise<AccountInfo>;
-
-  // returns the account balance in HBARs
-  getAccountBalance(): Promise<number>;
-
-  transferCrypto(options: {
-    transfers: SimpleTransfer[];
-    memo: string | null;
-    maxFee: number | null; // hbars
-    serviceFeesToPay: Record<string, number>;
-    serviceFeeToAddress: string | null;
-    onBeforeConfirm?: () => void;
-  }): Promise<TxReceipt>;
-
-  stakeHbar(options: {
-    nodeId: number | null;
-    accountId: string | null;
-  }): Promise<TxReceipt>;
-
-  approveAllowance(options: {
-    spenderAccountId: string;
-    amount: number;
-    assetType: string;
-    assetDetail?: ApproveAllowanceAssetDetail;
-  }): Promise<TxReceipt>;
-
-  deleteAllowance(options: {
-    assetType: string;
-    assetId: string;
-    spenderAccountId?: string;
-  }): Promise<TxReceipt>;
-
-  deleteAccount(options: { transferAccountId: string }): Promise<TxReceipt>;
-
-  associateTokens(options: { tokenIds: string[] }): Promise<TxReceipt>;
-
-  createToken(options: {
-    assetType: 'TOKEN' | 'NFT';
-    name: string;
-    symbol: string;
-    decimals: number;
-    supplyType: 'FINITE' | 'INFINITE';
-    initialSupply: number;
-    maxSupply: number;
-    expirationTime: string | undefined;
-    autoRenewAccountId: string;
-    tokenMemo: string;
-    freezeDefault: boolean;
-    kycPublicKey: string | undefined;
-    freezePublicKey: string | undefined;
-    pausePublicKey: string | undefined;
-    wipePublicKey: string | undefined;
-    supplyPublicKey: string | undefined;
-    feeSchedulePublicKey: string | undefined;
-    customFees: TokenCustomFee[] | undefined;
-  }): Promise<TxReceipt>;
 };
 
 export type MirrorStakingInfoServiceEndpoint = {
